+++ flmodules/Cargo.toml
@@ -44,13 +44,6 @@ reqwest = { version = "0.12", features = [
     "rustls-tls",
 ], default-features = false }
 bitflags = { version = "2", features = ["serde"] }
-x25519-dalek = {version = "2", features = ["serde"]}
-hex = "0.4.3"
-bincode = "1.3.3"
-sphinx-packet = { version = "0.2" }
-concurrent-queue = "2.5.0"
-rand_distr = "0.4.3"
 
 [dev-dependencies]
 env_logger = "0.11"
-flmodules = {path=".", features = ["testing"]}
