+++ flmodules/src/lib.rs
@@ -9,7 +9,6 @@ bitflags! {
         const ENABLE_PING = 0x8;
         const ENABLE_WEBPROXY = 0x10;
         const ENABLE_WEBPROXY_REQUESTS = 0x20;
-        const ENABLE_LOOPIX = 0x40;
     }
 }
 
@@ -22,4 +21,3 @@ pub mod ping;
 pub mod web_proxy;
 pub mod network;
 pub mod overlay;
-pub mod loopix;
