+++ flmodules/src/overlay/broker.rs
@@ -4,11 +4,59 @@ use flarch::{
     platform_async_trait,
 };
 
-use super::super::messages::{OverlayIn, OverlayInternal, OverlayMessage, OverlayOut};
+use super::messages::{OverlayIn, OverlayInternal, OverlayMessage, OverlayOut};
 use crate::{
     network::messages::{NetworkIn, NetworkMessage, NetworkOut},
     nodeconfig::NodeInfo,
+    random_connections::messages::{RandomIn, RandomMessage, RandomOut},
 };
+
+pub struct OverlayRandom {}
+
+impl OverlayRandom {
+    pub async fn start(
+        random: Broker<RandomMessage>,
+    ) -> Result<Broker<OverlayMessage>, BrokerError> {
+        let mut b = Broker::new();
+        // Translate RandomOut to OverlayOut, and OverlayIn to RandomIn.
+        // A module connected to the Broker<OverlayMessage> will get translations of the
+        // RandomOut messages, and can send messages to RandomIn using the Overlay.
+        b.link_bi(
+            random,
+            Box::new(|msg| {
+                if let RandomMessage::Output(out) = msg {
+                    let ret = match out {
+                        RandomOut::NodeIDsConnected(node_ids) => {
+                            OverlayOut::NodeIDsConnected(node_ids)
+                        }
+                        RandomOut::NodeInfosConnected(infos) => {
+                            OverlayOut::NodeInfosConnected(infos)
+                        }
+                        RandomOut::NetworkWrapperFromNetwork(id, module_message) => {
+                            OverlayOut::NetworkWrapperFromNetwork(id, module_message)
+                        }
+                        _ => return None,
+                    };
+                    return Some(ret.into());
+                }
+                None
+            }),
+            Box::new(|msg| {
+                if let OverlayMessage::Input(input) = msg {
+                    let ret = match input {
+                        OverlayIn::NetworkWrapperToNetwork(id, module_message) => {
+                            RandomIn::NetworkMapperToNetwork(id, module_message)
+                        }
+                    };
+                    return Some(RandomMessage::Input(ret));
+                }
+                None
+            }),
+        )
+        .await?;
+        Ok(b)
+    }
+}
 
 /**
  * Connects directly to the Network broker.
