--- flmodules/src/overlay/broker/mod.rs
@@ -1,3 +1,0 @@
-pub mod random;
-pub mod direct;
-pub mod loopix;
